requirements.txt
flask==2.0.1
python-docx==0.8.11
PyPDF2==2.10.5
scikit-learn==1.0.2
waitress==2.0.0
numpy>=1.19.2
pandas>=1.1.3
